Introduces type aliases.
